Moved asserting code out to StatefulPropCheckerAsserting enabler class, pathing way to support Future assertion.
