+++ jvm/core/src/main/scala/org/scalatest/prop/StatefulPropertyCheckModel.scala
@@ -6,6 +6,7 @@ import scala.compat.Platform.EOL
 import org.scalatest.{Assertion, Expectation, Fact}
 import org.scalatest.Assertions.succeed
 import org.scalatest.exceptions.{StackDepthException, GeneratorDrivenPropertyCheckFailedException}
+import org.scalatest.enablers.StatefulPropCheckerAsserting
 import org.scalatest.FailureMessages
 
 import org.scalactic.{source, Prettifier}
@@ -101,11 +102,6 @@ trait StatefulPropertyCheckModel[TCommand, TState, R] {
    */
   def postCondition(oldState: TState, newState: TState, command: TCommand, accCmd: IndexedSeq[TCommand], accRes: IndexedSeq[TState]): Boolean
 
-  private[scalatest] def indicateSuccess(message: => String, prettifier: Prettifier): R
-
-  private[scalatest] def indicateFailure(messageFun: StackDepthException => String, undecoratedMessage: => String, prettifier: Prettifier, optionalCause: Option[Throwable], pos: source.Position, 
-                                         initState: TState, initRnd: Randomizer, failingCmds: List[TCommand]): R
-
   case class NextRoseTree(value: Seq[TCommand]) extends RoseTree[Seq[TCommand]] {
     def shrinks: LazyListOrStream[RoseTree[Seq[TCommand]]] = {
       def resLazyListOrStream(theValue: Seq[TCommand]): LazyListOrStream[RoseTree[Seq[TCommand]]] = {
@@ -121,35 +117,6 @@ trait StatefulPropertyCheckModel[TCommand, TState, R] {
     }
   }
 
-  private def checkSut(szp: SizeParam, sut: SystemUnderTest[TCommand, TState], initState: TState, gen: Generator[TCommand], initRnd: Randomizer)(implicit pos: source.Position, prettifier: Prettifier): (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState], Option[TState]) = {
-
-    @tailrec def loop(count: Int, state: TState, rnd: Randomizer, accCmd: IndexedSeq[TCommand], accRnd: IndexedSeq[Randomizer], accRes: IndexedSeq[TState], failedPreconditionCount: Int): (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState], Option[TState]) = {
-      if (count > 0) {
-        val (cmd, newRnd) = command(state, gen, rnd)
-        if (preCondition(state, cmd, accCmd, accRes)) {
-          val newState = nextState(state, cmd)
-          val sutNewState = sut.nextState(state, cmd)
-          if (newState != sutNewState) 
-            (accCmd :+ cmd, accRnd :+ rnd, accRes :+ newState, Some(sutNewState))
-          else if (!postCondition(state, newState, cmd, accCmd, accRes)) 
-            (accCmd :+ cmd, accRnd :+ rnd, accRes :+ newState, Some(sutNewState))
-          else
-            loop(count - 1, newState, newRnd, accCmd :+ cmd, accRnd :+ newRnd, accRes :+ newState, 0)
-        }
-        else {
-          if (failedPreconditionCount < Generator.MaxLoopCount)
-            loop(count, state, newRnd, accCmd, accRnd, accRes, failedPreconditionCount + 1)
-          else
-            throw new IllegalStateException(FailureMessages.commandGeneratorExceededMaxLoopCount(prettifier, Generator.MaxLoopCount))  
-        }
-      }
-      else 
-        (accCmd, accRnd, accRes, None)
-    }
-
-    loop(szp.size, initState, initRnd, IndexedSeq.empty, IndexedSeq.empty, IndexedSeq.empty, 0)
-  }
-
   /**
    * Check the property with the given size parameter.
    *
@@ -158,116 +125,8 @@ trait StatefulPropertyCheckModel[TCommand, TState, R] {
    * @param prettifier the prettifier
    * @return the result of the property check
    */
-  def check(szp: SizeParam)(implicit pos: source.Position, prettifier: Prettifier): R = {
-
-    def tryRun(trySzp: SizeParam, initState: TState, gen: Generator[TCommand], initRnd: Randomizer): (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState], Option[TState]) = {
-      val sut = createSystemUnderTest(initState)
-      checkSut(trySzp, sut, initState, gen, initRnd)
-    }
-
-    def shrinkLoop(gen: Generator[TCommand], base: (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState], Option[TState]), remainings: (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState])): (IndexedSeq[TCommand], IndexedSeq[Randomizer], IndexedSeq[TState], Option[TState]) = {
-      val (remainingCmd, remainingRnd, remainingRes) = remainings
-      val (baseCmd, baseRnd, baseRes, baseFailingSutState) = base
-      if (remainingCmd.length > 2) {  
-        val halfSize = remainingCmd.length / 2
-
-        val firstHalfRemainingCmd = remainingCmd.take(halfSize)
-        val secondHalfRemainingCmd = remainingCmd.drop(halfSize)
-
-        val firstHalfRemainingRnd = remainingRnd.take(halfSize)
-        val secondHalfRemainingRnd = remainingRnd.drop(halfSize)
-
-        val firstHalfRemainingRes = remainingRes.take(halfSize)
-        val secondHalfRemainingRes = remainingRes.drop(halfSize)
-
-        val nextLength = PosZInt.ensuringValid(secondHalfRemainingRnd.length + 1)
-      
-        val (secondHalfCmd, secondHalfRnd, secondHalfRes, secondHalfFailingSutState) = tryRun(szp, firstHalfRemainingRes.last, gen, secondHalfRemainingRnd.head)
-        secondHalfFailingSutState match {
-          case Some(failingSutState) => // Successfully got failure using second half, we'll continue shrinking from there and forget about first half.
-            shrinkLoop(gen, (baseCmd, baseRnd, baseRes, secondHalfFailingSutState), (secondHalfRemainingCmd, secondHalfRemainingRnd, secondHalfRemainingRes))
-          case None => // Failed to get failure using second half, we'll continue with first half by using second half as base.
-            shrinkLoop(gen, (secondHalfCmd ++ baseCmd, secondHalfRnd ++ baseRnd, secondHalfRes ++ baseRes, baseFailingSutState), (firstHalfRemainingCmd, firstHalfRemainingRnd, firstHalfRemainingRes))
-        }
-      }
-      else 
-        (remainingCmd ++ baseCmd, remainingRnd ++ baseRnd, remainingRes ++ baseRes, baseFailingSutState)
-    }
-
-    val (initState, gen, initRnd) = initialize
-
-    val (firstAccCmd, firstAccRnd, firstAccRes, firstFailingSutState) = tryRun(szp, initState, gen, initRnd)
-    val (accCmd, accRnd, accRes, failingSutState) = 
-      if (firstFailingSutState.isEmpty) 
-        (firstAccCmd, firstAccRnd, firstAccRes, firstFailingSutState)
-      else 
-        shrinkLoop(gen, (IndexedSeq.empty, IndexedSeq.empty, IndexedSeq.empty, firstFailingSutState), (firstAccCmd, firstAccRnd, firstAccRes))
-
-    failingSutState match {
-      case Some(sutState) =>
-        val resInitState = accRes.head 
-        val resInitRnd = accRnd.head
-        val failureMsg = 
-          if (accRes.last == sutState)
-            FailureMessages.postConditionFailedAfterExecutingCommands + EOL + accCmd.map(_.toString).mkString("\n") + EOL + FailureMessages.initState(prettifier, resInitState) + EOL + FailureMessages.initSeed(prettifier, resInitRnd.seed)
-          else
-            FailureMessages.sutReturnedDifferentStateAfterExecutingCommands + EOL + accCmd.map(_.toString).mkString("\n") + EOL + FailureMessages.initState(prettifier, resInitState) + EOL + FailureMessages.initSeed(prettifier, resInitRnd.seed)
-        indicateFailure(sde => failureMsg, failureMsg, prettifier, None, pos, resInitState, resInitRnd, accCmd.toList)
-      case None =>
-        indicateSuccess(FailureMessages.propertyCheckSucceeded, prettifier)
-    }      
-  }
-
-}
-
-/**
- * Stateful property check model trait that can be used for building model for stateful property-based testing using assertions.
- *
- * This trait is used for ScalaTest assertions.
- */
-trait AssertiongStatefulPropertyCheckModel[TCommand, TState] extends StatefulPropertyCheckModel[TCommand, TState, Assertion] {
-
-  private[scalatest] def indicateSuccess(message: => String, prettifier: Prettifier): Assertion = succeed
-
-  private[scalatest] def indicateFailure(messageFun: StackDepthException => String, undecoratedMessage: => String, prettifier: Prettifier, optionalCause: Option[Throwable], pos: source.Position, 
-                                         initState: TState, initRnd: Randomizer, failingCmds: List[TCommand]): Assertion = 
-    throw new GeneratorDrivenPropertyCheckFailedException(
-      messageFun,
-      optionalCause,
-      pos,
-      Some((initState, initRnd)),
-      undecoratedMessage,
-      failingCmds,
-      None,
-      List.empty
-    )
-
-}
-
-/**
- * Stateful property check model trait that can be used for building model for stateful property-based testing using expectations.
- *
- * This trait is used for ScalaTest expectations.
- */
-trait ExpectationStatefulPropertyCheckModel[TCommand, TState] extends StatefulPropertyCheckModel[TCommand, TState, Expectation] {
-
-  private[scalatest] def indicateSuccess(message: => String, prettifier: Prettifier): Expectation = Fact.Yes(message, prettifier)
-
-  private[scalatest] def indicateFailure(messageFun: StackDepthException => String, undecoratedMessage: => String, prettifier: Prettifier, optionalCause: Option[Throwable], pos: source.Position, 
-                                         initState: TState, initRnd: Randomizer, failingCmds: List[TCommand]): Expectation = {
-    val gdpcfe =
-      new GeneratorDrivenPropertyCheckFailedException(
-        messageFun,
-        optionalCause,
-        pos,
-        Some((initState, initRnd)),
-        undecoratedMessage,
-        failingCmds,
-        None,
-        List.empty
-      )
-    val message: String = gdpcfe.getMessage
-    Fact.No(message, prettifier)  
+  def check(szp: SizeParam)(implicit asserting: StatefulPropCheckerAsserting[TCommand, TState, R], pos: source.Position, prettifier: Prettifier): R = {
+    asserting.check(this, szp, pos, prettifier)
   }
 
 }
